Addressed the three most common causes of issue #21
GRIP now detects when there's no world loaded and bails out of the
simulatestart, simulatestop, and simulatesingle functions before they
have a chance to segfault. There may be more out-of-order crashes
lurking deeper in the tree of possible actions, but now none of the
built-in buttosn will crash if they're hit with no world loaded.